* new MRC: made some modifications to the storage backend   - some refactoring was necessary to set the stage for restoring DB dumps   - the backend now uses direct BabuDB lookups for performance reasons   - ensured that 'stat' no longer returns negative access modes * MRC unit test now tests the new MRC
git-svn-id: https://xtreemfs.googlecode.com/svn/trunk@87 f8528748-c2c0-11dd-ba5c-d3b7c2b3b534

+++ servers/src/org/xtreemfs/new_mrc/ac/FileAccessManager.java
@@ -116,7 +116,7 @@ public class FileAccessManager {
         return getVolumeFileAccessPolicy(volumeId).translateAccessMode(accessMode);
     }
     
-    public long getPosixAccessMode(StorageManager sMan, FileMetadata file, String userId,
+    public short getPosixAccessMode(StorageManager sMan, FileMetadata file, String userId,
         List<String> groupIds) throws MRCException {
         return getVolumeFileAccessPolicy(sMan.getVolumeId()).getPosixAccessRights(sMan, file,
             userId, groupIds);
